refactor(audit): move snapshot detector assembly into its own package
The snapshot temporal-filter detector broke the pattern its peers follow:
make_fact_grounded_detectors and make_nullability_detectors each live in
their domain package (uniqueness/, nullability/) and the walker just
imports and composes them. The snapshot family instead had its assembly
(_make_snapshot_detectors) and a copy of dbt's default validity column
names living inside the orchestrator, so adding it had touched an
otherwise detector-agnostic file and duplicated a constant already
resolved in manifest/parse.py.

Give snapshot a peer package: src/dblect/snapshot owns both the pure
detector and make_snapshot_detectors(manifest). The walker now imports
nothing snapshot-specific; it composes makers like the others. The
duplicated default is replaced by a single DEFAULT_SNAPSHOT_VALIDITY_COLUMNS
exported from manifest/parse.py, kept as the fallback for snapshots whose
manifest carries no snapshot_meta_column_names block.

Behavior-preserving: full suite (766) and pyright/ruff clean.
