+++ src/dblect/manifest/__init__.py
@@ -4,6 +4,7 @@ from dblect.manifest.catalog import Catalog
 from dblect.manifest.dag import CycleError, Dag
 from dblect.manifest.parse import (
     DATA_FLOW_UID_PREFIXES,
+    DEFAULT_SNAPSHOT_VALIDITY_COLUMNS,
     Column,
     ConstraintSpec,
     ConstraintType,
@@ -17,6 +18,7 @@ from dblect.manifest.parse import (
 
 __all__ = [
     "DATA_FLOW_UID_PREFIXES",
+    "DEFAULT_SNAPSHOT_VALIDITY_COLUMNS",
     "Catalog",
     "Column",
     "ConstraintSpec",
